Rename: fibonacci.hpp => fibonacci.h
Let's use the same convention for filenames throughout.
